change logic for building filename
Now, if the built filename leads to a non-existing file, a new filename
will be generated with the current timestamp.
This is done to ensure files with a given file name are created at most
once. If they are deleted, new file name will be generated and used
until it is deleted or until the program is restarted/reconfigured.